+++ www/src/routes/items.js
@@ -125,6 +125,7 @@ router.get(["/details.html"], async function(req, res, next) {
     query ItemDetails($id: Int!) {
         getItemById(id: $id) {
             ... ItemAll
+            submittedBy
             getMob {
                 id
                 name
@@ -194,6 +195,7 @@ router.get(["/history.html"], async function(req, res, next) {
         getItemHistoryById(id: $id) {
             item {
                 ... ItemAll
+                submittedBy
                 getMob {
                     id
                     name
@@ -414,6 +416,11 @@ router.get(["/edit.html"], async function(req, res, next) {
 
     let item = data.getItemById;
     let itemStatCategories = data.getItemStatCategories;
+    if (item.official) {
+        const error = new Error("Official items cannot be edited for now.");
+        error.status = 403;
+        return next(error);
+    }
     let title = `Edit ${item.name}`;
     let vm = {
         item,